@@ -1,6 +1,6 @@
 package edu.usc.nlcaceres.workit.workit;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 public class MainActivity extends AppCompatActivity {
